reasoning_distillation/requirements.txt
torch
transformers
accelerate
datasets
peft
bitsandbytes
pillow
